refactor(reid): collapse slim modules into appearance and encoder
Drop the separate _lazy helper in favor of package __getattr__, merge
distance/extraction into appearance.py, and rename protocols to encoder.
