+++ src/trackers/core/reid/__init__.py
@@ -6,7 +6,7 @@
 
 from __future__ import annotations
 
-from trackers.core.reid._lazy import import_reid_symbol
+import importlib
 
 __all__ = [
     "FASTREID_MOT17_SBS50",
@@ -24,8 +24,12 @@ __all__ = [
     "resolve_model_card",
 ]
 
+REID_INSTALL_HINT = (
+    "ReID features require the optional `trackers[reid]` extra. Install with: pip install 'trackers[reid]'"
+)
+
 # NumPy-only symbols — safe to import without torch/timm/HF.
-from trackers.core.reid.distance import appearance_similarity
+from trackers.core.reid.appearance import appearance_similarity
 from trackers.core.reid.eval.datasets import ReIDSplit, load_market1501, load_msmt17
 from trackers.core.reid.eval.metrics import ReIDMetrics, compute_reid_metrics
 from trackers.core.reid.feature_bank import FeatureBank
@@ -40,10 +44,19 @@ _LAZY_EXPORTS: dict[str, tuple[str, str]] = {
 }
 
 
+def _import_reid_symbol(module_name: str, attr_name: str) -> object:
+    """Import a heavy ReID symbol, rewriting missing-extra errors."""
+    try:
+        module = importlib.import_module(module_name)
+        return getattr(module, attr_name)
+    except ImportError as exc:
+        raise ImportError(REID_INSTALL_HINT) from exc
+
+
 def __getattr__(name: str) -> object:
     if name in _LAZY_EXPORTS:
         module_name, attr_name = _LAZY_EXPORTS[name]
-        value = import_reid_symbol(module_name, attr_name)
+        value = _import_reid_symbol(module_name, attr_name)
         globals()[name] = value
         return value
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
